+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 - Add currency symbols
 - Add `as` and `into` keywords for unit conversion
+- Fix some units not being simplified
 
 ## 4.1.0 - 2026 Jul 29
 - Add currency units